Create check-mail
This checker is for one purpose, in an ecommerce when we receive an email with a new purchase, we make a copy of a specific google sheet based on the product purchased and make the buyer an editor for that sheet. after that we send him an email with some instructions.